Our first difference comparison!

+++ play-version6/app/models/IwpStepDifference.scala
@@ -2,8 +2,8 @@ package models
 
 
 case class IwpStepDifference ( step: Int,
-                               notEqual: Map[String, (Double,Double)],
-                               equal: Map[String, Double],
+                               notEqual: Map[String, (BigDecimal, BigDecimal)],
+                               equal: Map[String, BigDecimal],
                                leftEmpty: Boolean,
                                leftMissing: Seq[String],
                                rightEmpty: Boolean,
